타입 정리
